# PERF: further improve take (reindex/unstack) for ArrayManager