--- a/pandas/core/array_algos/take.py
+++ b/pandas/core/array_algos/take.py
@@ -149,7 +149,7 @@ def take_1d(
         return arr.take(indexer, fill_value=fill_value, allow_fill=allow_fill)
 
     if not allow_fill:
-        return np.take(arr, indexer)
+        return arr.take(indexer)
 
     indexer, dtype, fill_value, mask_info = _take_preprocess_indexer_and_fill_value(
         arr, indexer, None, fill_value, allow_fill