--- a/pandas/core/array_algos/take.py
+++ b/pandas/core/array_algos/take.py
@@ -534,13 +534,8 @@ def _take_preprocess_indexer_and_fill_value(
         if dtype != arr.dtype and (out is None or out.dtype != dtype):
             # check if promotion is actually required based on indexer
             mask = indexer == -1
-            # error: Item "bool" of "Union[Any, bool]" has no attribute "any"
-            # [union-attr]
-            needs_masking = mask.any()  # type: ignore[union-attr]
-            # error: Incompatible types in assignment (expression has type
-            # "Tuple[Union[Any, bool], Any]", variable has type
-            # "Optional[Tuple[None, bool]]")
-            mask_info = mask, needs_masking  # type: ignore[assignment]
+            needs_masking = mask.any()
+            mask_info = mask, needs_masking
             if needs_masking:
                 if out is not None and out.dtype != dtype:
                     raise TypeError("Incompatible type for fill_value")