@jbrockmendel 2021-03-12
Can/should we just have take_nd do an earlier check for ndim == 1 and then dispatch to take_1d?

In NDArrayBackedExtensionArray.take if i change take_nd to

```
        if self.ndim == 1 and axis == 0 and allow_fill:
            indices = ensure_platform_int(indices)
            validate_indices(indices, len(self))
            new_data = take_1d(
                self._ndarray, indices, allow_fill=allow_fill, fill_value=fill_value
            )

        else:
             new_data = [what we do right now]
```

then on the time_full_product benchmark referenced https://github.com/pandas-dev/pandas/issues/39146#issuecomment-792817698 we shave off 25-30% relative to master (non-AM)