@jorisvandenbossche 2021-03-12
> im fine with that, just suggesting if we have a clear optimization for the general take lets use it

Certainly, see https://github.com/pandas-dev/pandas/pull/40405